requirements_python13.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
sentence-transformers==2.6.1
huggingface_hub==0.20.3
#transformers==4.38.2
transformers==4.39.3
tokenizers==0.15.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.2
auth0-python==4.10.0
av==15.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
cfgv==3.5.0
charset-normalizer==3.4.2
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confluent-kafka==2.11.1
crewai==0.126.0
cryptography==45.0.7
ctranslate2==4.6.0
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.0
faiss-cpu==1.13.2
faster-whisper==1.2.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
Flask-Mail==0.9.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gevent==25.9.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-cloud-texttospeech==2.33.0
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
humanfriendly==10.0
identify==2.6.15
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
ipykernel==6.30.0
ipython==8.18.1
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
Js2Py==0.74
json5==0.12.0
json_repair==0.50.1
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.23
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.8
lark==1.2.2
litellm==1.68.0
llvmlite==0.44.0
lxml==6.0.0
mailjet-rest==1.5.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.2.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.10.0
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.75.0
openai-whisper==20250625
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.1
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pickle-mixin==1.0.2
pillow==11.3.0
platformdirs==4.3.8
posthog==5.4.0
pre_commit==4.5.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyjsparser==2.7.1
PyJWT==2.10.1
pyparsing==3.3.1
pypdfium2==4.30.0
PyPika==0.48.9
PyPrind==2.11.3
pyproject_hooks==1.2.0
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-json-logger==3.3.0
pyttsx3==2.99
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
Send2Trash==1.8.3
#sentence-transformers==2.2.2
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.42
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
#tokenizers==0.22.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
types-python-dateutil==2.9.0.20250708
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.16
uvicorn==0.35.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.2.1
yarl==1.20.1
zipp==3.23.0
zope.event==6.1
zope.interface==8.1.1
zstandard==0.23.0
